+++ Neolution.Extensions.Identity.Abstractions/Options/GoogleOptions.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.Extensions.Identity.Abstractions.Options
+{
+    /// <summary>
+    /// Options for Google OpenID Connect.
+    /// </summary>
+    public class GoogleOptions
+    {
+        /// <summary>
+        /// Gets or sets the client identifier.
+        /// </summary>
+        public string? ClientId { get; set; }
+    }
+}