+++ Neolution.Extensions.Identity.Abstractions/Options/NeolutionIdentityOptions.cs
@@ -1,0 +1,18 @@
+﻿namespace Neolution.Extensions.Identity.Abstractions.Options
+{
+    /// <summary>
+    /// Neolution Identity Options
+    /// </summary>
+    public class NeolutionIdentityOptions
+    {
+        /// <summary>
+        /// Gets or sets the Google options.
+        /// </summary>
+        public GoogleOptions Google { get; set; } = new();
+
+        /// <summary>
+        /// Gets or sets the password hasher options.
+        /// </summary>
+        public PasswordHasherOptions PasswordHasher { get; set; } = new();
+    }
+}